Organize format container classes slightly differently
