+++ src/DspAdpcm2/Containers/Bxstm/BxstmCodec.cs
@@ -1,4 +1,4 @@
-﻿namespace DspAdpcm.Containers.Structures
+﻿namespace DspAdpcm.Containers.Bxstm
 {
     /// <summary>
     /// The different audio codecs used in BRSTM, BCSTM, and BFSTM files.
